Show skeleton when loading search
Also slightly refactors it and splits search into multiple components
