+++ frontend/src/components/application/ApplicationCard.stories.tsx
@@ -1,20 +1,34 @@
-import React from "react"
-import { Meta } from "@storybook/react"
-import ApplicationCard from "./ApplicationCard"
+import { ApplicationCard } from "./ApplicationCard"
 import { faker } from "@faker-js/faker"
 
-export default {
-  title: "Components/Application/ApplicationCard",
+import type { Meta, StoryObj } from "@storybook/react"
+
+const meta: Meta<typeof ApplicationCard> = {
   component: ApplicationCard,
-} as Meta<typeof ApplicationCard>
+}
 
-export const Generated = () => {
-  const application = {
-    id: faker.datatype.uuid(),
-    icon: faker.image.image(),
-    name: faker.commerce.product(),
-    summary: faker.commerce.productDescription(),
-  }
+export default meta
+type Story = StoryObj<typeof ApplicationCard>
 
-  return <ApplicationCard application={application} />
+export const Primary: Story = {
+  args: {
+    application: {
+      id: faker.datatype.uuid(),
+      icon: faker.image.image(),
+      name: faker.commerce.product(),
+      summary: faker.commerce.productDescription(),
+    },
+  },
+}
+
+export const inACard: Story = {
+  args: {
+    application: {
+      id: faker.datatype.uuid(),
+      icon: faker.image.image(),
+      name: faker.commerce.product(),
+      summary: faker.commerce.productDescription(),
+    },
+    inACard: true,
+  },
 }
